% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCompensation2FluoChannelNames}
\alias{addCompensation2FluoChannelNames}
\title{add compensation to fluo column names}
\usage{
addCompensation2FluoChannelNames(ff)
}
\arguments{
\item{ff}{a flowCore::flowFrame}
}
\value{
new flowCore::flowFrame with the new column names
}
\description{
: technical utility to add "Comp-" prefix to all
column names for fluorochrom channels
}
