% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{appendCellID}
\alias{appendCellID}
\title{append 'Original_ID' column to a flowframe}
\usage{
appendCellID(ff, eventIDs = seq_len(flowCore::nrow(ff)))
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{eventIDs}{an integer vector containing the values to be added in
as Original ID's}
}
\value{
new flowCore::flowFrame containing the added 'Original_ID' column
}
\description{
: on a flowCore::flowFrame, append a 'Original_ID' column.
This column can be used in plots comparing the events pre and post gating.
If the 'Original_ID' column already exists, the function does nothing
}
\examples{

data(OMIP021Samples)

retFF <- appendCellID(OMIP021Samples[[1]])

}
