% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{applyScaleTransforms}
\alias{applyScaleTransforms}
\title{apply scale transforms}
\usage{
applyScaleTransforms(ff, transList, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{transList}{a flowCore::transformList}

\item{...}{other arguments (not used)}
}
\value{
the transformed flowFrame
}
\description{
apply scale transforms
}
