% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{areFluoCols}
\alias{areFluoCols}
\title{find flow frame columns that represent fluorochrome channel}
\usage{
areFluoCols(
  x,
  toRemovePatterns = c("FSC", "SSC", "Time", "Original_ID", "File", "SampleID")
)
}
\arguments{
\item{x}{a flowCore::flowFrame or a flowCore::flowSet}

\item{toRemovePatterns}{a vector of string patterns that are to be considered
as non fluorochrome}
}
\value{
a vector of booleans of which the dimension is equal to the number of
columns in ff
}
\description{
: find flow frame columns that represent fluorochrome channel
}
\examples{

data(OMIP021Samples)

areFluoCols(OMIP021Samples)

}
