% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{areSignalCols}
\alias{areSignalCols}
\title{find flow frame columns that represent true signal}
\usage{
areSignalCols(
  ff,
  toRemovePatterns = c("Time", "Original_ID", "File", "SampleID")
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{toRemovePatterns}{a vector of string patterns that are to be considered
as non signal}
}
\value{
a vector of booleans of which the dimension is equal to the number of
columns in ff
}
\description{
: find flow frame columns that represent true signal
}
