% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R, R/CytoPipeline.R
\docType{class}
\name{experimentName}
\alias{experimentName}
\alias{experimentName<-}
\alias{sampleFiles}
\alias{sampleFiles<-}
\alias{addProcessingStep}
\alias{removeProcessingStep}
\alias{getNbProcessingSteps}
\alias{getProcessingStep}
\alias{getProcessingStepNames}
\alias{cleanProcessingSteps}
\alias{showProcessingSteps}
\alias{execute}
\alias{buildCytoPipelineFromCache}
\alias{checkCytoPipelineConsistencyWithCache}
\alias{as.list.CytoPipeline}
\alias{export2JSONFile}
\alias{getCytoPipelineExperimentNames}
\alias{getCytoPipelineObjectFromCache}
\alias{getCytoPipelineObjectInfos}
\alias{getCytoPipelineFlowFrame}
\alias{getCytoPipelineScaleTransform}
\alias{plotCytoPipelineProcessingQueue}
\alias{deleteCytoPipelineCache}
\alias{CytoPipeline-class}
\alias{CytoPipeline-class,}
\alias{CytoPipeline}
\alias{show,CytoPipeline-method}
\alias{CytoPipeline,missing-method}
\alias{CytoPipeline,list-method}
\alias{CytoPipeline,character-method}
\title{CytoPipeline class}
\usage{
experimentName(x)

experimentName(x) <- value

sampleFiles(x)

sampleFiles(x) <- value

addProcessingStep(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  newPS
)

removeProcessingStep(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  index
)

getNbProcessingSteps(x, whichQueue = c("scale transform", "pre-processing"))

getProcessingStep(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  index
)

getProcessingStepNames(x, whichQueue = c("scale transform", "pre-processing"))

cleanProcessingSteps(
  x,
  whichQueue = c("both", "scale transform", "pre-processing")
)

showProcessingSteps(x, whichQueue = c("scale transform", "pre-processing"))

execute(x, path = ".", rmCache = FALSE)

buildCytoPipelineFromCache(experimentName, path = ".")

checkCytoPipelineConsistencyWithCache(x, path = ".")

\method{as.list}{CytoPipeline}(x, ...)

export2JSONFile(x, path)

getCytoPipelineExperimentNames(
  path = ".",
  pattern = NULL,
  ignore.case = FALSE,
  fixed = FALSE
)

getCytoPipelineObjectFromCache(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  sampleFile = NULL,
  objectName,
  path = "."
)

getCytoPipelineObjectInfos(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  sampleFile = NULL,
  objectName,
  path = "."
)

getCytoPipelineFlowFrame(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  sampleFile,
  objectName,
  path = "."
)

getCytoPipelineScaleTransform(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  objectName,
  path = "."
)

plotCytoPipelineProcessingQueue(
  x,
  whichQueue = c("pre-processing", "scale transform"),
  sampleFile = NULL,
  path = ".",
  title = TRUE,
  box.type = "ellipse",
  lwd = 1,
  box.prop = 0.5,
  box.cex = 0.7,
  cex.txt = 0.7,
  box.size = 0.1,
  dtext = 0.15,
  ...
)

deleteCytoPipelineCache(x, path = ".")

\S4method{show}{CytoPipeline}(object)

\S4method{CytoPipeline}{missing}(
  object,
  experimentName = "default_experiment",
  sampleFiles = character()
)

\S4method{CytoPipeline}{list}(object)

\S4method{CytoPipeline}{character}(object)
}
\arguments{
\item{x}{CytoPipeline object}

\item{path}{base path, a subdirectory with name equal to the experiment will
be created to store the output data, in particular the experiment cache}

\item{rmCache}{if TRUE, starts by removing the already existing cache
directory corresponding to the experiment}
}
\description{
Class representing a flow cytometry pipeline, and composed of two processing
queues, i.e. lists of CytoProcessingStep objects :
\itemize{
\item a list of CytoProcessingStep(s) for pre-calculation of scale transformations
per channel
\item a list of CytoProcessingStep(s) for the pre-processing of flow frames
}
}
\section{Slots}{

\describe{
\item{\code{scaleTransformProcessingQueue}}{A \code{list()} of CytoProcessingStep objects
containing the steps for obtaining the scale transformations per channel}

\item{\code{flowFramesPreProcessingQueue}}{A \code{list()} of CytoProcessingStep objects
containing the steps for pre-processing of the samples flow frames}

\item{\code{sampleFiles}}{A \code{character()} vector storing all fcs files to be
run into the pipeline}

\item{\code{savePreprocessedFiles}}{if TRUE, will save pre-processed fcs file in
/QC subdirectory}

\item{\code{savePlotsInFiles}}{if TRUE, will save files corresponding to default
generated plot at each pre-processing step}

\item{\code{saveScaleTransform}}{if TRUE, will save rds object storing scale
transformation list generated}

\item{\code{scaleTransformFile}}{basename of the file to use to save the scale
transformation list (if 'saveScaleTransform' == TRUE)}
}}

