% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{execute}
\alias{execute}
\title{executing CytoPipeline object}
\usage{
execute(x, path = ".", rmCache = FALSE)
}
\arguments{
\item{x}{CytoPipeline object}

\item{path}{base path, a subdirectory with name equal to the experiment will
be created to store the output data, in particular the experiment cache}

\item{rmCache}{if TRUE, starts by removing the already existing cache
directory corresponding to the experiment}
}
\value{
nothing
}
\description{
this function triggers the execution of the processing queues of
a CytoPipeline object.
First, the scale tranform processing queue is run, taking
the set of sample names as an implicit first input. At the end of the queue,
a scale transform List is assumed to be created.
Second, the flowFrame pre-processing queue, reapeatedly for each sample file.
The scale transform list generated in the previous step is taken as implicit
input, together with the initial sample file. At the end of the queue run, a
pre-processed flowFrame is assumed to be generated.
No change is made on the input CytoPipeline object, all results are stored in
the cache.
}
