% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findTimeChannel}
\alias{findTimeChannel}
\title{find time channel in flowSet/flowFrame}
\usage{
findTimeChannel(obj, excludeChannels = c())
}
\arguments{
\item{obj}{a flowCore::flowFrame or flowCore::flowSet}

\item{excludeChannels}{vector of column names to exclude in the search}
}
\value{
a character, name of the found channel that should be representing
time. If not found, returns NULL.
}
\description{
tries to find a channel in a flowSet/flowFrame that could
be the time channel. First tries to identify a channel name containing the
'time' string, then tries to identify a single monotonically increasing channel.
}
