% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChannelNamesFromMarkers}
\alias{getChannelNamesFromMarkers}
\title{get channel names from markers}
\usage{
getChannelNamesFromMarkers(ff, markers)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{markers}{a vector of markers, either provided as :
\itemize{
\item an array of booleans (referring to flowFrame columns)
\item an array of integers (indices in flowFrame columns)
\item an array of characters (exact markers or channel patterns)
}}
}
\value{
a character vector, containing the names of the corresponding
channels
}
\description{
finds name of channels corresponding to user provided markers
}
