% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{getCytoPipelineExperimentNames}
\alias{getCytoPipelineExperimentNames}
\title{Find CytoPipeline experiments stored in a file cache}
\usage{
getCytoPipelineExperimentNames(
  path = ".",
  pattern = NULL,
  ignore.case = FALSE,
  fixed = FALSE
)
}
\arguments{
\item{path}{root path to locate the search for file caches}

\item{pattern}{optional pattern limiting the search for experiment
names}

\item{ignore.case}{(TRUE/FALSE) used in pattern matching (grepl)}

\item{fixed}{(TRUE/FALSE) used in pattern matching (grepl)}
}
\value{
a vector of character containing found experiment names
}
\description{
This function looks into a path for stored file caches
and gets the corresponding experiment names
}
