% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{getCytoPipelineObjectFromCache}
\alias{getCytoPipelineObjectFromCache}
\title{Retrieves an object from a file cache}
\usage{
getCytoPipelineObjectFromCache(
  x,
  path = ".",
  whichQueue = c("scale transform", "pre-processing"),
  sampleFile = NULL,
  objectName
)
}
\arguments{
\item{x}{a CytoPipeline object}

\item{path}{root path to locate the search for file caches}

\item{whichQueue}{which queue to look into}

\item{sampleFile}{which sampleFile is looked for:
\itemize{
\item if whichQueue == "scale transform", the sampleFile is ignored
\item if NULL and whichQueue == "pre-processing", the sampleFile is
defaulted to the first one belonging to the experiment
}}

\item{objectName}{(character) which object name to look for}
}
\value{
the found object (or stops with an error message if the
target object is not found)
}
\description{
Given a CytoPipeline object, this function retrieves
a specific object in the corresponding file cache
}
