% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{getCytoPipelineObjectInfos}
\alias{getCytoPipelineObjectInfos}
\title{File cache objects information}
\usage{
getCytoPipelineObjectInfos(
  x,
  path = ".",
  whichQueue = c("scale transform", "pre-processing"),
  sampleFile = NULL
)
}
\arguments{
\item{x}{a CytoPipeline object}

\item{path}{root path to locate the search for file caches}

\item{whichQueue}{which queue to look into}

\item{sampleFile}{which sampleFile is looked for:
\itemize{
\item if whichQueue == "scale transform", the sampleFile is ignored
\item if NULL and whichQueue == "pre-processing", the sampleFile is
defaulted to the first one belonging to the experiment
}}
}
\value{
a dataframe with the collected information about the
found objects (or stops with an error message if no target object
was found)
}
\description{
Given a CytoPipeline object, this function retrieves
the information related to a specific object name, i.e.
object name and object class
}
\details{
internal function for the time being
}
