% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{getCytoPipelineScaleTransform}
\alias{getCytoPipelineScaleTransform}
\title{Retrieves a transformList object from a file cache}
\usage{
getCytoPipelineScaleTransform(
  x,
  path = ".",
  whichQueue = c("scale transform", "pre-processing"),
  sampleFile = NULL,
  objectName
)
}
\arguments{
\item{x}{a CytoPipeline object}

\item{path}{root path to locate the search for file caches}

\item{whichQueue}{which queue to look into}

\item{sampleFile}{which sampleFile is looked for:
\itemize{
\item if whichQueue == "scale transform", the sampleFile is ignored
\item if NULL and whichQueue == "pre-processing", the sampleFile is
defaulted to the first one belonging to the experiment
}}

\item{objectName}{(character) which object name to look for}
}
\value{
the found flowFrame (or stops with an error message if the
target object is not found, or if the object is no transformList)
}
\description{
Given a CytoPipeline object, this function retrieves
a specific flowCore::transformList object in the corresponding
file cache
}
