% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFCSFileName}
\alias{getFCSFileName}
\title{get fcs file name}
\usage{
getFCSFileName(ff)
}
\arguments{
\item{ff}{a flowCore::flowFrame}
}
\value{
the basename of $FILENAME keyword
}
\description{
get basename of $FILENAME keyword if exists
}
\examples{

data(OMIP021Samples)

fName <- getFCSFileName(OMIP021Samples[[1]])
}
