% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplotEvents}
\alias{ggplotEvents}
\title{plot events in 1D or 2D, using ggplot2}
\usage{
ggplotEvents(
  obj,
  xChannel,
  yChannel = NULL,
  nDisplayCells = Inf,
  seed = NULL,
  bins = 216,
  fill = "lightblue",
  alpha = 0.2,
  xScale = c("linear", "logicle"),
  yScale = c("linear", "logicle"),
  xLogicleParams = list(w = 2, m = 6.42, a = 0, t = 262144),
  yLogicleParams = list(w = 2, m = 6.42, a = 0, t = 262144),
  xLinearRange = NULL,
  yLinearRange = NULL,
  transList = NULL,
  runTransforms = FALSE
)
}
\arguments{
\item{obj}{a flowCore::flowFrame or flowCore::flowSet}

\item{xChannel}{channel (name or index) or marker name to be displayed
on x axis}

\item{yChannel}{channel (name or index) or marker name to be displayed
on y axis}

\item{nDisplayCells}{maximum number of events that will be plotted. If
the number of events exceed this number, a sub-sampling will be performed}

\item{seed}{seed used for sub-sampling (if any)}

\item{bins}{used in geom_hex}

\item{fill}{used in geom_density}

\item{alpha}{used in geom_density}

\item{xScale}{scale to be used for the x axis}

\item{yScale}{scale to be used for the y axis}

\item{xLogicleParams}{if (xScale == "logicle"), the parameters of the logicle
transformation to be used, as a list(w = ..., m = ..., a = ..., t = ...)}

\item{yLogicleParams}{if (yScale == "logicle"), the parameters of the logicle
transformation to be used, as a list(w = ..., m = ..., a = ..., t = ...)}

\item{xLinearRange}{if (xScale == "linear"), the x axis range to be used}

\item{yLinearRange}{if (yScale == "linear"), the y axis range to be used}

\item{transList}{optional list of scale transformations to be applied to each
channel. If it is non null, 'x/yScale', 'x/yLogicleParams' and
'x/yLinear_range' will be discarded.}

\item{runTransforms}{(TRUE/FALSE) only taken into account if transList is
not NULL. Will 'transList' result in data being effectively transformed ?
\itemize{
\item If TRUE, than the data will undergo transformations prior to
visualization.
\item If FALSE, the axis will be scaled but the data themselves are
not transformed.
}}
}
\value{
a list of ggplot objects
}
\description{
plot events of specific channels of either :
flowCore::flowFrame, or flowCore::flowSet
in 2D or 1D, mimicking FlowJo type of graph. \cr
if 1D : geom_density will be used \cr
if 2D : geom_hex will be used \cr
}
