% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplotFilterEvents}
\alias{ggplotFilterEvents}
\title{plot filtered events in 2D, using ggplot}
\usage{
ggplotFilterEvents(
  ffPre,
  ffPost,
  xChannel,
  yChannel,
  nDisplayCells = 10000,
  seed = NULL,
  size = 0.5,
  xScale = c("linear", "logicle"),
  yScale = c("linear", "logicle"),
  xLogicleParams = list(w = 2, m = 6.42, a = 0, t = 262144),
  yLogicleParams = list(w = 2, m = 6.42, a = 0, t = 262144),
  xLinearRange = NULL,
  yLinearRange = NULL,
  interactive = FALSE
)
}
\arguments{
\item{ffPre}{a flowCore::flowFrame, before applying filter}

\item{ffPost}{a flowCore::flowFrame, after applying filter}

\item{xChannel}{channel (name or index) or marker name to be displayed on x axis}

\item{yChannel}{channel (name or index) or marker name to be displayed on y axis}

\item{nDisplayCells}{maximum number of events that will be plotted. If
the number of events exceed this number, a subsampling will be performed}

\item{seed}{seed used for sub-sampling (if any)}

\item{size}{used by geom_point()}

\item{xScale}{scale to be used for the x axis}

\item{yScale}{scale to be used for the y axis}

\item{xLogicleParams}{if (xScale == "logicle"), the parameters of the logicle
transformation to be used, as a list(w = ..., m = ..., a = ..., t = ...)}

\item{yLogicleParams}{if (yScale == "logicle"), the parameters of the logicle
transformation to be used, as a list(w = ..., m = ..., a = ..., t = ...)}

\item{xLinearRange}{if (xScale == "linear"), linear range to be used}

\item{yLinearRange}{if (yScale == "linear"), linear range to be used}

\item{interactive}{if TRUE, transform the scaling formats such that the
ggcyto::x_scale_logicle() and ggcyto::y_scale_logicle() do work with
plotly::ggplotly()}
}
\value{
a ggplot object
}
\description{
plot events of specific channels of either :
flowCore::flowFrame, or
flowCore::flowSet
in 2D, showing the impact of applying a filter between :
\itemize{
\item a 'pre' flowframe
}
}
