% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{handlingProcessingSteps}
\alias{handlingProcessingSteps}
\alias{addProcessingStep}
\alias{removeProcessingStep}
\alias{getNbProcessingSteps}
\alias{getProcessingStep}
\alias{getProcessingStepNames}
\alias{cleanProcessingSteps}
\alias{showProcessingSteps}
\title{handling processing steps in CytoPipeline objects}
\usage{
addProcessingStep(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  newPS
)

removeProcessingStep(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  index
)

getNbProcessingSteps(x, whichQueue = c("scale transform", "pre-processing"))

getProcessingStep(
  x,
  whichQueue = c("scale transform", "pre-processing"),
  index
)

getProcessingStepNames(x, whichQueue = c("scale transform", "pre-processing"))

cleanProcessingSteps(
  x,
  whichQueue = c("both", "scale transform", "pre-processing")
)

showProcessingSteps(x, whichQueue = c("scale transform", "pre-processing"))
}
\arguments{
\item{x}{a CytoPipeline object}

\item{whichQueue}{selects the processing queue for which we manage the
processing
steps}

\item{newPS}{the new processing step to be added (CytoProcessingStep object)}

\item{index}{index of the processing step to remove}
}
\value{
\itemize{
\item for \code{addProcessingStep}: the updated CytoPipeline object
}

\itemize{
\item for \code{removeProcessingStep}: the updated CytoPipeline object
}

\itemize{
\item for \code{getNbProcessingSteps}: the number of processing steps present
in the target queue
}

\itemize{
\item for \code{getProcessingStep}: the obtained CytoProcessingStep object
}

\itemize{
\item for \code{getProcessingStepNames}: the vector of step names
}

\itemize{
\item for \code{cleanProcessingSteps}: the updated CytoPipeline object
}

\itemize{
\item for \code{showProcessingSteps}: nothing (only console display side
effect is required)
}
}
\description{
functions to manipulate processing steps in processing queues
of CytoPipeline objects
}
\section{Functions}{
\itemize{
\item \code{addProcessingStep}: adds a processing step in one of the
processing queues (at the end), returns the modified CytoPipeline object

\item \code{removeProcessingStep}: removes a processing step from one of
the processing queues, returns the modified CytoPipeline object

\item \code{getNbProcessingSteps}: gets the number of processing
steps in a processing queue

\item \code{getProcessingStep}: gets a processing step at a
specific index of a processing queue

\item \code{getProcessingStepNames}: gets a character vector of all
processing step names of a specific processing queue

\item \code{cleanProcessingSteps}: deletes all processing steps in one
or both processing queues, returns the modified CytoPipeline object

\item \code{showProcessingSteps}: shows all processing steps in a
processing queue
}}

