% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OMIP021Samples}
\alias{OMIP021Samples}
\title{OMIP021Samples dataset}
\format{
a flowCore::flowSet with two different flowFrames
each one contains one flow cytometry sample
corresponding to Donor1.fcs and Donor2.fcs in following source.
A subsampling of 5,000 events has been performed on each file.
}
\source{
\url{https://flowrepository.org/experiments/305}
}
\description{
OMIP021Samples dataset
}
\keyword{data}
