% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{plotCytoPipelineProcessingQueue}
\alias{plotCytoPipelineProcessingQueue}
\title{Plots a processing queue}
\usage{
plotCytoPipelineProcessingQueue(
  x,
  whichQueue = c("pre-processing", "scale transform"),
  sampleFile = NULL,
  path = ".",
  title = TRUE,
  box.type = "ellipse",
  lwd = 1,
  box.prop = 0.5,
  box.cex = 0.7,
  cex.txt = 0.7,
  box.size = 0.1,
  dtext = 0.15,
  ...
)
}
\arguments{
\item{x}{a CytoPipeline object}

\item{whichQueue}{which queue to look into}

\item{sampleFile}{which sampleFile is looked for:
\itemize{
\item if whichQueue == "scale transform", the sampleFile is ignored
\item if NULL and whichQueue == "pre-processing", the sampleFile is
defaulted to the first one belonging to the experiment
}}

\item{path}{root path to locate the search for file caches}

\item{title}{if TRUE, adds a title to the plot}

\item{box.type}{shape of label box (rect, ellipse, diamond,
round, hexa, multi)}

\item{lwd}{default line width of arrow and box (one numeric value)}

\item{box.prop}{length/width ratio of label box (one numeric value)}

\item{box.cex}{relative size of text in boxes (one numeric value)}

\item{cex.txt}{relative size of arrow text (one numeric value)}

\item{box.size}{size of label box (one numeric value)}

\item{dtext}{controls the position of arrow text relative to arrowhead
(one numeric value)}

\item{...}{other arguments passed to diagram::plotmat()}
}
\value{
nothing
}
\description{
This functions displays a plot of a processing
queue of a CytoPipeline object, using diagram::plotmat().
\itemize{
\item If a step is in run state for all sample files, the
corresponding box appears in green
\item If a step is in non run state for at least one sample file,
the corresponding box appears in orange
\item If at least one step is not consistent with cache, the whole
set of boxes appears in red
}
}
