% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{qualityControlFlowClean}
\alias{qualityControlFlowClean}
\title{perform QC with flowCut}
\usage{
qualityControlFlowClean(
  ff,
  preTransform = FALSE,
  transList = NULL,
  outputDiagnostic = FALSE,
  outputDir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{preTransform}{if TRUE, apply the transList scale transform prior to
running the gating algorithm}

\item{transList}{applied in conjunction with preTransform}

\item{outputDiagnostic}{if TRUE, stores diagnostic files generated by
flowClean in outputDir directory}

\item{outputDir}{used in conjunction with outputDiagnostic}

\item{verbose}{if TRUE messages comments on the QC process}

\item{...}{additional parameters passed to flowClean::clean()}
}
\value{
a flowCore::flowFrame with removed low quality events from the input
}
\description{
this function is a wrapper around flowClean::clean()
function.
It also pre-selects the channels to be handled (=> all signal channels)
}
