% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{readSampleFiles}
\alias{readSampleFiles}
\title{Read fcs sample files}
\usage{
readSampleFiles(sampleFiles, whichSamples = "all", ...)
}
\arguments{
\item{sampleFiles}{a vector of character path to sample files}

\item{whichSamples}{either 'all' if all sample files need to be read, or
a vector of indexes pointing to the sampleFiles vector}

\item{...}{additional parameters passed to flowCore file reading functions.}
}
\value{
either a flowCore::flowSet or a flowCore::flowFrame if
length(sampleFiles) == 1
}
\description{
Wrapper around flowCore::read.fcs() or flowCore::read.flowSet().
Also adds a "Cell_ID" additional column, used in flowFrames comparison
}
