% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeChannels}
\alias{removeChannels}
\title{remove channels from a flowFrame}
\usage{
removeChannels(ff, channels)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{channels}{the channel names to be removed}
}
\value{
a new flowCore::flowFrame with the removed channels
}
\description{
: in a flowCore::flowFrame, remove the channels of the given
names.
}
\examples{

data(OMIP021Samples)

retFF <- removeChannels(OMIP021Samples[[1]],
                        channel = "FSC-A")

}
