% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDeadCellsDeGate}
\alias{removeDeadCellsDeGate}
\title{remove dead cells from a flowFrame}
\usage{
removeDeadCellsDeGate(
  ff,
  preTransform = FALSE,
  transList = NULL,
  LDMarker,
  keepPositivePop = FALSE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{preTransform}{if TRUE, apply the transList scale transform prior to
running the gating algorithm}

\item{transList}{applied in conjunction with preTransform == TRUE}

\item{LDMarker}{a character containing the exact name of the marker
corresponding to Live/Dead channel, or the Live/Dead channel name itself}

\item{keepPositivePop}{logical flag stating whether we want to keep,
after gating, the population that is positive for \code{LDMarker} (TRUE)
or negative for \code{LDmarker} (FALSE)}

\item{...}{additional parameters passed to flowDensity::deGate()}
}
\value{
a flowCore::flowFrame with removed dead cells from the input
}
\description{
this function removes dead cells from a flowFrame, using a
specific '(a)live/dead' channel, and the flowDensity::deGate() gating
function (see doc of the flowDensity package)
}
\examples{

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

suppressWarnings(ff_m <- removeMarginsPeacoQC(x = fsRaw[[2]]))
    
ff_c <-
    compensateFromMatrix(ff_m,
                         matrixSource = "fcs")        

transList <- 
    estimateScaleTransforms(        
        ff = ff_c,
        fluoMethod = "estimateLogicle",
        scatterMethod = "linear",
        scatterRefMarker = "BV785 - CD3")

ff_lcells <-
    removeDeadCellsDeGate(ff_c,
                          preTransform = TRUE,
                          transList = transList,
                          LDMarker = "L/D Aqua - Viability",
                          keepPositivePop = FALSE)
                            
}
