% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDeadCellsGateTail}
\alias{removeDeadCellsGateTail}
\title{remove dead cells from a flowFrame}
\usage{
removeDeadCellsGateTail(
  ff,
  preTransform = FALSE,
  transList = NULL,
  LDMarker,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{preTransform}{if TRUE, apply the transList scale transform prior to
running the gating algorithm}

\item{transList}{applied in conjunction with preTransform == TRUE}

\item{LDMarker}{a character containing the exact name of the marker
corresponding to Live/Dead channel}

\item{...}{additional parameters passed to openCyto::gate_tail()}
}
\value{
a flowCore::flowFrame with removed dead cells from the input
}
\description{
this function removes dead cells from a flowFrame, using a
specific '(a)live/dead' channel, and the openCyto::gate_tail() gating
function (see doc of the openCyto package)
}
