% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDeadCellsManual}
\alias{removeDeadCellsManual}
\title{remove dead cells from a flowFrame using manual gating}
\usage{
removeDeadCellsManual(
  ff,
  preTransform = FALSE,
  transList = NULL,
  FSCChannel,
  LDMarker,
  gateData,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{preTransform}{boolean, if TRUE: the transList list of scale transforms
will be applied first on the LD channel.}

\item{transList}{applied in conjunction with preTransform == TRUE}

\item{FSCChannel}{a character containing the exact name of the forward
scatter channel}

\item{LDMarker}{a character containing the exact name of the marker
corresponding to (a)Live/Dead channel}

\item{gateData}{a numerical vector containing the polygon gate coordinates
first the FSC-A channel coordinates of each points of the polygon gate,
then the LD channel coordinates of each points (prior to scale transfom)}

\item{...}{additional parameters passed to flowCore::polygonGate()}
}
\value{
a flowCore::flowFrame with removed dead cells from the input
}
\description{
remove dead cells from a flowFrame, using manual gating in the
FSC-A, '(a)Live/Dead' 2D representation. The function uses
flowCore::polygonGate()
}
