% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDebrisFlowClustTmix}
\alias{removeDebrisFlowClustTmix}
\title{remove debris from a flowFrame, using flowClust}
\usage{
removeDebrisFlowClustTmix(ff, FSCChannel, SSCChannel, nClust, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{FSCChannel}{the name of the FSC channel}

\item{SSCChannel}{the name of the SSC channel}

\item{nClust}{number of clusters to identify}

\item{...}{additional parameters passed to flowClust::tmixFilter()}
}
\value{
a flowCore::flowFrame with removed debris events from the input
}
\description{
this function removes debris from a flowFrame,
using clustering capabilities of flowClust::tmixFilter(). The idea is to
pre-select a number of clusters to be found in the (FSC,SSC) 2D view, and
eliminate the cluster that is the closest to the origin.
}
