% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDebrisManualGate}
\alias{removeDebrisManualGate}
\title{remove debris from a flowFrame using manual gating}
\usage{
removeDebrisManualGate(ff, FSCChannel, SSCChannel, gateData, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{FSCChannel}{a character containing the exact name of the forward
scatter channel}

\item{SSCChannel}{a character containing the exact name of the side scatter
channel}

\item{gateData}{a numerical vector containing the polygon gate coordinates
first the FSC-A channel coordinates of each points of the polygon gate,
then the SSC-A channel coordinates of each points.}

\item{...}{additional parameters passed to flowCore::polygonGate()}
}
\value{
a flowCore::flowFrame with removed debris events from the input
}
\description{
remove debris from a flowFrame, using manual gating in the
FSC-A, SSC-A 2D representation. The function internally uses
flowCore::polygonGate()
}
