% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDoubletsFlowStats}
\alias{removeDoubletsFlowStats}
\title{remove doublets from a flowFrame, using flowStats}
\usage{
removeDoubletsFlowStats(
  ff,
  areaChannels,
  heightChannels,
  widerGate = FALSE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{areaChannels}{a character vector containing the name of the 'area type'
channels one wants to use}

\item{heightChannels}{a character vector containing the name of the
'height type' channels one wants to use}

\item{widerGate}{a boolean as wider_gate parameter to
flowStats::singletGate()}

\item{...}{additional parameters passed to flowStats::singletGate()}
}
\value{
a flowCore::flowFrame with removed doublets events from the input
}
\description{
Wrapper around flowStats::singletGate().
Can apply the flowStats function subsequently on several channel pairs,
e.g. (FSC-A, FSC-H) and (SSC-A, SSC-H)
}
