% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDoubletsPeacoQC}
\alias{removeDoubletsPeacoQC}
\title{remove doublets from a flowFrame, using PeacoQC}
\usage{
removeDoubletsPeacoQC(
  ff,
  areaChannels,
  heightChannels,
  nmads = rep(4, length(areaChannels)),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{areaChannels}{a character vector containing the name of the 'area type'
channels one wants to use}

\item{heightChannels}{a character vector containing the #' name of the
'height type' channels one wants to use}

\item{nmads}{a numeric vector with the bandwidth above the ratio allowed, per
channels pair (cells are kept if the ratio between -A channel[i] and
-H channel[i] is smaller than the median ratio + nmad[i] times the median
absolute deviation of the ratios). Default is 4, for all channel pairs.}

\item{verbose}{If set to TRUE, the median ratio and width will be printed.}

\item{...}{additional parameters passed to PeacoQC::RemoveDoublets()}
}
\value{
a flowCore::flowFrame with removed doublets events from the input
}
\description{
wrapper around PeacoQC::RemoveDoublets().
Can apply the PeacoQC function subsequently on several channel pairs,
e.g. (FSC-A, FSC-H) and (SSC-A, SSC-H)
}
