% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{selectRandomSamples}
\alias{selectRandomSamples}
\title{select a nb of sample files randomly}
\usage{
selectRandomSamples(sampleFiles, nSamples, seed = NULL, ...)
}
\arguments{
\item{sampleFiles}{a vector of character path to sample files}

\item{nSamples}{number of samples to randomly select. If \code{nSamples} is
higher than nb of available samples, the output will be all samples}

\item{seed}{an optional seed parameters (provided to ease reproducibility).}

\item{...}{additional parameters passed (not used here).}
}
\value{
a subset of \code{sampleFiles}, randomly selected
}
\description{
Wrapper around sample(sampleFiles, nSamples).
For the needs of making it a CytoProcessingStep:
\itemize{
\item takes \code{sampleFiles} as an explicit first parameter
\item can manage a seed
\item allow passing ... as a parameter (not used)
}
}
\examples{

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

nRandomSamples <- 1
selectSampleFiles <- selectRandomSamples(sampleFiles, 
                                         nSamples = nRandomSamples,
                                         seed = 1)
                                         
}
