% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsample}
\alias{subsample}
\title{sub-sampling of a flowFrame}
\usage{
subsample(ff, nSamples, seed = NULL)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{nSamples}{number of samples to be obtained using sub-sampling}

\item{seed}{can be set for reproducibility of event sub-sampling}
}
\value{
new flowCore::flowFrame with the obtained subset of samples
}
\description{
: sub-samples a flowFrame
with the specified number of samples, without replacement.
adds also a column 'Original_ID' if not already present in flowFrame.
}
