% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{applyScaleTransforms}
\alias{applyScaleTransforms}
\title{apply scale transforms}
\usage{
applyScaleTransforms(ff, transList, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{transList}{a flowCore::transformList}

\item{...}{other arguments (not used)}
}
\value{
the transformed flowFrame
}
\description{
wrapper around flowCore::transform() that discards any
additional parameter passed in (...)
}
\examples{

data(OMIP021Samples)

transListPath <- file.path(system.file("extdata", 
                                       package = "CytoPipeline"),
                           "OMIP021_TransList.rds") 

transList <- readRDSObject(transListPath)

ff_c <- compensateFromMatrix(OMIP021Samples[[1]],
                             matrixSource = "fcs")  

ff_t <- applyScaleTransforms(ff_c, transList = transList)

}
