% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline.R
\docType{class}
\name{CytoPipeline-class}
\alias{CytoPipeline-class}
\alias{CytoPipeline-class,}
\alias{CytoPipeline}
\alias{show,CytoPipeline-method}
\alias{CytoPipeline,missing-method}
\alias{CytoPipeline,list-method}
\alias{CytoPipeline,character-method}
\alias{as.list.CytoPipeline}
\alias{experimentName}
\alias{experimentName<-}
\alias{sampleFiles}
\alias{sampleFiles<-}
\title{CytoPipeline class}
\usage{
\S4method{show}{CytoPipeline}(object)

\S4method{CytoPipeline}{missing}(
  object,
  experimentName = "default_experiment",
  sampleFiles = character()
)

\S4method{CytoPipeline}{list}(object)

\S4method{CytoPipeline}{character}(object)

\method{as.list}{CytoPipeline}(x, ...)

experimentName(x)

experimentName(x) <- value

sampleFiles(x)

sampleFiles(x) <- value
}
\arguments{
\item{object}{a \code{character()} containing a JSON input}

\item{experimentName}{the experiment name}

\item{sampleFiles}{the sample files}

\item{x}{a \code{CytoPipeline} object}

\item{...}{additional arguments (not used here)}

\item{value}{the new value to be assigned}
}
\value{
\itemize{
\item for \code{as.list.CytoPipeline}: the obtained list
}
}
\description{
Class representing a flow cytometry pipeline, and composed of two processing
queues, i.e. lists of CytoProcessingStep objects :
\itemize{
\item a list of CytoProcessingStep(s) for pre-calculation of scale
transformations per channel
\item a list of CytoProcessingStep(s) for the pre-processing of flow frames
}
}
\section{Slots}{

\describe{
\item{\code{scaleTransformProcessingQueue}}{A \code{list()} of CytoProcessingStep objects
containing the steps for obtaining the scale transformations per channel}

\item{\code{flowFramesPreProcessingQueue}}{A \code{list()} of CytoProcessingStep objects
containing the steps for pre-processing of the samples flow frames}

\item{\code{experimentName}}{A \code{character()} containing the experiment (run) name}

\item{\code{sampleFiles}}{A \code{character()} vector storing all fcs files to be
run into the pipeline}

\item{\code{saveScaleTransform}}{if TRUE, will save rds object storing scale}

\item{\code{scaleTransformFile}}{basename of the file to use to save the scale}
}}

