% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipeline-functions.R
\name{exportCytoPipeline}
\alias{exportCytoPipeline}
\alias{export2JSONFile}
\title{exporting CytoPipeline objects}
\usage{
export2JSONFile(x, path)
}
\arguments{
\item{x}{a CytoPipeline object}

\item{path}{the full path to the name of the file to be created}
}
\value{
\itemize{
\item for \code{export2JSONFile}: nothing
}
}
\description{
functions to export CytoPipeline objects in various
formats
}
\section{Functions}{
\itemize{
\item \code{export2JSONFile}: exports a CytoPipeline object
to a JSON file (writing the file = side effect)
}}

