% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{getAcquiredCompensationMatrix}
\alias{getAcquiredCompensationMatrix}
\title{extract compensation matrix from a flowCore::flowFrame}
\usage{
getAcquiredCompensationMatrix(ff)
}
\arguments{
\item{ff}{a flowCore::flowFrame}
}
\value{
the found compensation matrix
}
\description{
helper function retrieving the compensation matrix stored
in fcs file (if any). It scans the following keywords: $SPILL, $spillover and
$SPILLOVER
}
