% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTransfoParams}
\alias{getTransfoParams}
\title{get tranformation parameters for a specific channel}
\usage{
getTransfoParams(transList, channel)
}
\arguments{
\item{transList}{a flowCore::transformList}

\item{channel}{channel name}
}
\value{
If the transformation exists for the specified channel, and is either
recognized as a logicle transfo or a linear transfo, a list with two slots:
\itemize{
\item $type a character containing the transfo type ('logicle' or 'linear')
\item $params_list a list of named numeric, according to transfo type
}

Otherwise, NULL is returned.
}
\description{
investigates a flowCore::tranformList object to get the type
and parameters of the transformation applying to a specific channel
}
