% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{qualityControlFlowClean}
\alias{qualityControlFlowClean}
\title{perform QC with flowClean}
\usage{
qualityControlFlowClean(
  ff,
  preTransform = FALSE,
  transList = NULL,
  outputDiagnostic = FALSE,
  outputDir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{preTransform}{if TRUE, apply the transList scale transform prior to
running the gating algorithm}

\item{transList}{applied in conjunction with preTransform}

\item{outputDiagnostic}{if TRUE, stores diagnostic files generated by
flowClean in outputDir directory}

\item{outputDir}{used in conjunction with outputDiagnostic}

\item{verbose}{if TRUE messages comments on the QC process}

\item{...}{additional parameters passed to flowClean::clean()}
}
\value{
a flowCore::flowFrame with removed low quality events from the input
}
\description{
this function is a wrapper around flowClean::clean()
function.
It also pre-selects the channels to be handled (=> all signal channels)
}
\examples{

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

ff_QualityControl <- 
    qualityControlFlowClean(fsRaw[[2]],
                            binSize = 0.01, # default
                            nCellCutoff = 500, # default
                            cutoff = "median", # default
                            fcMax = 1.3, # default
                            nstable = 5)

}
