% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{qualityControlFlowCut}
\alias{qualityControlFlowCut}
\title{perform QC with flowCut}
\usage{
qualityControlFlowCut(
  ff,
  preTransform = FALSE,
  transList = NULL,
  outputDiagnostic = FALSE,
  outputDir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{preTransform}{if TRUE, apply the transList scale transform prior to
running the gating algorithm}

\item{transList}{applied in conjunction with preTransform}

\item{outputDiagnostic}{if TRUE, stores diagnostic files generated by
flowCut in outputDir directory}

\item{outputDir}{used in conjunction with outputDiagnostic}

\item{verbose}{if TRUE messages comments on the QC process}

\item{...}{additional parameters passed to flowCut::flowCut()}
}
\value{
a flowCore::flowFrame with removed low quality events from the input
}
\description{
this function is a wrapper around flowCut::flowCut()
function.
It also pre-selects the channels to be handled (=> all signal channels)
}
\examples{

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

ff_QualityControl <-
    qualityControlFlowCut(
        fsRaw[[2]],
        MaxContin = 0.1,
        MeanOfMeans = 0.13,
        MaxOfMeans = 0.15,
        MaxValleyHgt = 0.1,
        MaxPercCut = 0.3,
        LowDensityRemoval = 0.1,
        RemoveMultiSD = 7,
        AlwaysClean = FALSE,
        IgnoreMonotonic = FALSE,
        MonotonicFix = NULL,
        Measures = c(1:8))
      
}
