% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{readSampleFiles}
\alias{readSampleFiles}
\title{Read fcs sample files}
\usage{
readSampleFiles(
  sampleFiles,
  whichSamples = "all",
  channelMarkerFile = NULL,
  ...
)
}
\arguments{
\item{sampleFiles}{a vector of character path to sample files}

\item{whichSamples}{either 'all' if all sample files need to be read, or
a vector of indexes pointing to the sampleFiles vector}

\item{channelMarkerFile}{an optional path to a csv file which provides the
mapping between channels and markers. If provided, this csv file should
contain a \code{Channel} column, and a \code{Marker} column. Optionally a 'Used'
column can be provided as well (TRUE/FALSE). Channels for which the 'Used'
column is set to FALSE will not be incorporated in the created flowFrame.}

\item{...}{additional parameters passed to flowCore file reading functions.}
}
\value{
either a flowCore::flowSet or a flowCore::flowFrame if
length(sampleFiles) == 1
}
\description{
Wrapper around flowCore::read.fcs() or flowCore::read.flowSet().
Also adds a "Cell_ID" additional column, used in flowFrames comparison
}
\examples{

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
res <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

#res

# create a flowCore::flowFrame with one single sample
res2 <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = 2,
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

#res2

}
