% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDoubletsFlowStats}
\alias{removeDoubletsFlowStats}
\title{remove doublets from a flowFrame, using flowStats}
\usage{
removeDoubletsFlowStats(
  ff,
  areaChannels,
  heightChannels,
  widerGate = FALSE,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{areaChannels}{a character vector containing the name of the 'area type'
channels one wants to use}

\item{heightChannels}{a character vector containing the name of the
'height type' channels one wants to use}

\item{widerGate}{a boolean as wider_gate parameter to
flowStats::singletGate()}

\item{...}{additional parameters passed to flowStats::singletGate()}
}
\value{
a flowCore::flowFrame with removed doublets events from the input
}
\description{
Wrapper around flowStats::singletGate().
Can apply the flowStats function subsequently on several channel pairs,
e.g. (FSC-A, FSC-H) and (SSC-A, SSC-H)
}
\examples{

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

suppressWarnings(ff_m <- removeMarginsPeacoQC(x = fsRaw[[2]]))
    
ff_c <-
    compensateFromMatrix(ff_m,
                         matrixSource = "fcs")        

ff_s <-
    removeDoubletsFlowStats(ff_c,
                            areaChannels = c("FSC-A", "SSC-A"),
                            heightChannels = c("FSC-H", "SSC-H"),
                            widerGate = TRUE)
                            
}
