% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeMarginsPeacoQC}
\alias{removeMarginsPeacoQC}
\title{remove margin events using PeacoQC}
\usage{
removeMarginsPeacoQC(x, ...)
}
\arguments{
\item{x}{a flowCore::flowSet or a flowCore::flowFrame}

\item{...}{additional parameters passed to PeacoQC::RemoveMargins()}
}
\value{
either a flowCore::flowSet or a flowCore::flowFrame depending on
the input.
}
\description{
Wrapper around PeacoQC::RemoveMargins().
Also pre-selects the channels to be handled (=> all signal channels)
If input is a flowSet, it applies removeMargins() to each flowFrame of the
flowSet.
}
\examples{

rawDataDir <- 
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <- 
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL
fsRaw <- readSampleFiles(sampleFiles, 
                         truncate_max_range = truncateMaxRange,
                         min.limit = minLimit)
suppressWarnings(ff_m <- removeMarginsPeacoQC(x = fsRaw[[2]]))
ggplotFilterEvents(ffPre = fsRaw[[2]],
                   ffPost = ff_m,
                   xChannel = "FSC-A",
                   yChannel = "SSC-A")
}
