% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{writeFlowFrame}
\alias{writeFlowFrame}
\title{write flowFrame to disk}
\usage{
writeFlowFrame(ff, fileName, format = c("fcs", "csv"), ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{...}{other arguments (not used)}

\item{transList}{a flowCore::transformList}
}
\value{
nothing
}
\description{
wrapper around flowCore::write.FCS() or utils::write.csv
that discards any additional parameter passed in (...)
}
\examples{

transListPath <- paste0(system.file("extdata", 
                                    package = "CytoPipeline"),
                        "/OMIP021_TransList.rds") 

transList <- readRDSObject(transListPath)

ff_c <- compensateFromMatrix(OMIP021Samples[[1]],
                             matrixSource = "fcs")  

ff_t <- applyScaleTransforms(ff_c, transList = transList)


rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
res <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)
    
ff_c <- compensateFromMatrix(res[[2]], matrixSource = "fcs") 
outputDir <- withr::local_tempdir()
outputFile <- paste0(outputDir, "/fcs_export.csv")
writeFlowFrame(ff_c, fileName = outputFile, format = "csv")

}
